refactor : split `FixedCostList` and `FixedCostTotalPrice` components
- split two components to show user streaming UI
- use Suspense and Fallback UI
