+++ src/components/expenseTracker/FixedCostTotalPrice.tsx
@@ -1,0 +1,15 @@
+import { useSuspenseQuery } from '@tanstack/react-query';
+import { ExpenseTracker, getFixedCostPaymentsByMonth } from '../../supabase';
+import { currentMonth, monetizeWithSeparator } from '../../utils';
+import { queryKey } from '../../constants';
+
+const FixedCostTotalPrice = () => {
+	const { data } = useSuspenseQuery<ExpenseTracker[]>({
+		queryKey: [...queryKey.EXPENSE_TRACKER, 'fixedCost'],
+		queryFn: () => getFixedCostPaymentsByMonth(currentMonth - 1),
+	});
+
+	return <p aria-label="total price">{monetizeWithSeparator(data.reduce((acc, curr) => acc + +curr.priceIntegerPart, 0))}</p>;
+};
+
+export default FixedCostTotalPrice;
