+++ src/components/expenseTracker/loader/FixedCostTotalPriceLoader.tsx
@@ -1,0 +1,7 @@
+import { SkeletonLoader } from '../../common';
+
+const FixedCostTotalPriceLoader = () => {
+	return <SkeletonLoader width={'100%'} height={'36px'} theme={'light'} />;
+};
+
+export default FixedCostTotalPriceLoader;
